+++ internal/modelcontext/handles.go
@@ -93,69 +93,3 @@ func (t *HandleTable) DecodeKnownText(value string) string {
 	}
 	return value
 }
-
-// HandleStreamDecoder restores HandleTable values without leaking a handle
-// split across provider chunks.
-type HandleStreamDecoder struct {
-	table   *HandleTable
-	pending string
-}
-
-func NewHandleStreamDecoder(table *HandleTable) *HandleStreamDecoder {
-	return &HandleStreamDecoder{table: table}
-}
-
-func (d *HandleStreamDecoder) Feed(chunk string) string {
-	if d == nil || d.table == nil {
-		return chunk
-	}
-	combined := d.pending + chunk
-	d.pending = ""
-	if combined == "" {
-		return ""
-	}
-	start := len(combined)
-	for start > 0 && isHandleTokenByte(combined[start-1]) {
-		start--
-	}
-	tail := combined[start:]
-	// The prefix is immutable after construction, so no lock is needed.
-	prefix := d.table.table.prefix
-	if couldBeNumericHandle(tail, prefix) {
-		d.pending = tail
-		combined = combined[:start]
-	}
-	return d.table.DecodeKnownText(combined)
-}
-
-func (d *HandleStreamDecoder) Flush() string {
-	if d == nil || d.table == nil {
-		return ""
-	}
-	tail := d.table.DecodeKnownText(d.pending)
-	d.pending = ""
-	return tail
-}
-
-func isHandleTokenByte(value byte) bool {
-	return value >= 'a' && value <= 'z' || value >= 'A' && value <= 'Z' ||
-		value >= '0' && value <= '9' || value == '-' || value == '_'
-}
-
-func couldBeNumericHandle(value, prefix string) bool {
-	if value == "" || prefix == "" {
-		return false
-	}
-	if strings.HasPrefix(prefix, value) {
-		return true
-	}
-	if !strings.HasPrefix(value, prefix) || len(value) == len(prefix) {
-		return false
-	}
-	for _, char := range value[len(prefix):] {
-		if char < '0' || char > '9' {
-			return false
-		}
-	}
-	return true
-}
